requirements.txt
blinker==1.6.2
blis==0.7.9
catalogue==2.0.8
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
cymem==2.0.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
filelock==3.12.2
Flask==2.2.3
Flask-Mail==0.9.1
fsspec==2023.1.0
googletrans==4.0.0rc1
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.16.4
hyperframe==5.2.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
MarkupSafe==2.1.2
mpmath==1.3.0
murmurhash==1.0.9
networkx==2.6.3
nltk==3.7
numpy==1.21.6
packaging==23.1
pandas==1.3.5
pathy==0.10.1
Pillow==9.5.0
preshed==3.0.8
pydantic==1.9.2
pyspellchecker==0.7.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
regex==2023.5.5
requests==2.31.0
rfc3986==1.5.0
safetensors==0.3.2
scikit-learn==1.0.2
scipy==1.7.3
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
spacy==3.4.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
sympy==1.10.1
thinc==8.1.10
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.65.0
transformers==4.30.2
typer==0.4.2
urllib3<2.0
wasabi==0.10.1
Werkzeug==2.2.3
typing-extensions==4.4.0